requirements.txt
# Django core
django>=5.1,<6.0
django-environ>=0.11.2

# Database
psycopg2-binary>=2.9.9

# Channels (WebSocket)
channels>=4.1.0
channels-redis>=4.2.0
daphne>=4.1.2

# Celery (filas e tasks assíncronas)
celery>=5.3.6
redis>=5.0.3
django-celery-beat>=2.7.0
django-celery-results>=2.5.1

# OpenAI
openai>=1.30.1

# HTTP client (chamadas UazAPI)
httpx>=0.27.0

# Frontend helpers
django-htmx>=1.19.0
whitenoise>=6.7.0

# Auth
django-allauth>=0.63.3

# Utilities
pillow>=10.4.0
python-slugify>=8.0.4
django-extensions>=3.2.3
django-crispy-forms>=2.1
crispy-tailwind>=1.0.3
django-cors-headers>=4.4.0

# Dev
ipython>=8.24.0
django-debug-toolbar>=4.4.2
